refactor(node): consolidate numpy_reader and numpy_writer into numpy_file module
