+++ docs/api/nodes.md
@@ -31,7 +31,7 @@ live module docstrings over historical inventories.
 
 ### NumPy Readers
 
-::: cuvis_ai.node.numpy_reader
+::: cuvis_ai.node.numpy_file.NpyReader
     options:
       show_root_heading: true
       heading_level: 3
@@ -145,7 +145,7 @@ live module docstrings over historical inventories.
 
 ### NumPy Writers
 
-::: cuvis_ai.node.numpy_writer
+::: cuvis_ai.node.numpy_file.NumpyFeatureWriterNode
     options:
       show_root_heading: true
       heading_level: 3
